Corrigindo implementação de trafficLights, refatoração do delay para o enum e correção do metodo getColor
