+++ java-para-iniciante/src/exercs/cap12/trafficsimulator/ColorEnum.java
@@ -1,6 +1,14 @@
 package exercs.cap12.trafficsimulator;
 
 enum ColorEnum {
-    RED, GREEN, YELLOW;
+
+        RED(8000), GREEN(8000), YELLOW(3000);
+
+        int delay;
+
+        ColorEnum(int delay) {
+            this.delay = delay;
+        }
+
 
 }
